Tidied up

+++ source/ui/ui_service.cpp
@@ -82,8 +82,8 @@ namespace ui
         else if (p_message->get_type() == messages::message_window_resize::TYPE)
         {
             auto p_window_resize_message = static_cast<messages::message_window_resize*>(p_message);
-            m_ui_width = p_window_resize_message->get_width();
-            m_ui_height = p_window_resize_message->get_height();
+            m_ui_width = static_cast<float>(p_window_resize_message->get_width());
+            m_ui_height = static_cast<float>(p_window_resize_message->get_height());
 
             if (mp_root_node)
             {
