+++ source/ui/ui_service.hpp
@@ -41,9 +41,9 @@ namespace ui
         /// @brief The root UI node
         std::unique_ptr<ui::root> mp_root_node{nullptr};
         /// @brief The UI width
-        int m_ui_width{0};
+        float m_ui_width{0.0f};
         /// @brief The UI height
-        int m_ui_height{0};
+        float m_ui_height{0.0f};
     };
 } /// namespace ui
